6-26

+++ mountainsortalg/detect_events.h
@@ -19,6 +19,6 @@
 #include <QVector>
 #include "mlcompute.h"
 
-QVector<double> detect_events(const QVector<double>& X, double detect_threshold, double detect_interval, int sign);
+QVector<bigint> detect_events(const QVector<double>& X, double detect_threshold, double detect_interval, int sign);
 
 #endif // DETECT_EVENTS_H
